backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-socketio==5.10.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
supabase==2.0.2
httpx==0.25.2
pydantic==2.5.0
pydantic-settings==2.1.0
python-socketio[asyncio_client]==5.10.0
aiofiles==23.2.1 